+++ android/src/main/java/com/hoxfon/react/RNTwilioVoice/TwilioVoiceModule.java
@@ -20,7 +20,6 @@ import android.os.PowerManager;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.content.LocalBroadcastManager;
-import android.telephony.TelephonyManager;
 import android.util.Log;
 import android.view.Window;
 import android.view.WindowManager;
@@ -57,6 +56,7 @@ import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_DID_RECEI
 import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_NOT_READY;
 import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_READY;
 
+
 public class TwilioVoiceModule extends ReactContextBaseJavaModule implements ActivityEventListener, LifecycleEventListener {
 
     public static String TAG = "RNTwilioVoice";
@@ -68,14 +68,10 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
 
     private boolean isReceiverRegistered = false;
     private VoiceBroadcastReceiver voiceBroadcastReceiver;
-    private PhoneCallBroadcastReceiver phoneCallBroadcastReceiver;
 
     // Empty HashMap, contains parameters for the Outbound call
     private HashMap<String, String> twiMLParams = new HashMap<>();
     
-    public static final String EVENT_PHONE_CALL_STARTED = "phoneCallStarted";
-    public static final String EVENT_PHONE_CALL_ENDED = "phoneCallEnded";
-
     public static final String INCOMING_CALL_INVITE          = "INCOMING_CALL_INVITE";
     public static final String INCOMING_CALL_NOTIFICATION_ID = "INCOMING_CALL_NOTIFICATION_ID";
     public static final String NOTIFICATION_TYPE             = "NOTIFICATION_TYPE";
@@ -100,6 +96,7 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
     private NotificationManager notificationManager;
     private CallNotificationManager callNotificationManager;
     private ProximityManager proximityManager;
+    private PhoneCallManager phoneCallManager;
 
     private String accessToken;
 
@@ -138,6 +135,7 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         callNotificationManager = new CallNotificationManager();
         proximityManager = new ProximityManager(reactContext, eventManager);
         headsetManager = new HeadsetManager(eventManager);
+        phoneCallManager = new PhoneCallManager(reactContext, eventManager);
 
         notificationManager = (android.app.NotificationManager) reactContext.getSystemService(Context.NOTIFICATION_SERVICE);
 
@@ -146,7 +144,6 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
          * or incoming call messages in this Activity.
          */
         voiceBroadcastReceiver = new VoiceBroadcastReceiver();
-        phoneCallBroadcastReceiver = new PhoneCallBroadcastReceiver();
         registerReceiver();
 
         TwilioVoiceModule.callNotificationMap = new HashMap<>();
@@ -340,11 +337,6 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
 
             registerActionReceiver();
 
-            IntentFilter phoneCallIntentFilter = new IntentFilter();
-            phoneCallIntentFilter.addAction(TelephonyManager.ACTION_PHONE_STATE_CHANGED);
-            getReactApplicationContext().registerReceiver(
-                    phoneCallBroadcastReceiver, phoneCallIntentFilter);
-
             isReceiverRegistered = true;
         }
     }
@@ -519,61 +511,6 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                 sharedPrefEditor.commit();
             } else {
                 Log.e(TAG, "received broadcast unhandled action " + action);
-            }
-        }
-    }
-
-    private class PhoneCallBroadcastReceiver extends BroadcastReceiver {
-
-        private int lastState = TelephonyManager.CALL_STATE_IDLE;
-
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            if (BuildConfig.DEBUG) {
-                Log.d(TAG, "received onReceive action 1 " + intent.toString());
-            }
-            String stateStr = intent.getExtras().getString(TelephonyManager.EXTRA_STATE);
-            int state = 0;
-            if(stateStr.equals(TelephonyManager.EXTRA_STATE_IDLE)){
-                state = TelephonyManager.CALL_STATE_IDLE;
-            }
-            else if(stateStr.equals(TelephonyManager.EXTRA_STATE_OFFHOOK)){
-                state = TelephonyManager.CALL_STATE_OFFHOOK;
-            }
-            else if(stateStr.equals(TelephonyManager.EXTRA_STATE_RINGING)){
-                state = TelephonyManager.CALL_STATE_RINGING;
-            }
-
-            onCallStateChanged(context, state);
-            lastState = state;
-        }
-
-        public void onCallStateChanged(Context context, int state) {
-            if (BuildConfig.DEBUG) {
-                Log.d(TAG, "received onCallStateChanged action " + state + " lastState " + lastState);
-            }
-            if (lastState == state) {
-                //No change, debounce extras
-                return;
-            }
-            else {
-                switch (state)
-                    {
-                        case TelephonyManager.CALL_STATE_IDLE:
-                            if (lastState == TelephonyManager.CALL_STATE_OFFHOOK) {
-                                eventManager.sendEvent(EVENT_PHONE_CALL_ENDED, null);
-                            }
-                            break;
-                        case TelephonyManager.CALL_STATE_RINGING:
-                            break;
-                        case TelephonyManager.CALL_STATE_OFFHOOK:
-                            if (lastState == TelephonyManager.CALL_STATE_RINGING) {
-                                eventManager.sendEvent(EVENT_PHONE_CALL_STARTED, null);
-                            }
-                            break;
-                        default:
-                            break;
-                    }
             }
         }
     }
